provider/requirements.txt
fastapi
uvicorn
sqlalchemy
pydantic
joblib
numpy
python-dotenv
pymysql
scikit-learn
python-multipart